Cover toon output across CLI surface
